requirements.txt
absl-py==0.15.0
astor==0.8.1
audioread==2.1.9
backports.functools-lru-cache==1.6.4
backports.weakref==1.0rc1
bleach==1.5.0
cffi==1.15.0
chardet==4.0.0
cloudpickle==1.3.0
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.30
decorator==4.4.2
easydict==1.9
enum34==1.1.10
funcsigs==1.0.2
future==0.18.2
futures==3.3.0
gast==0.5.3
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.41.1
h5py==2.10.0
html5lib==0.9999999
idna==2.10
imageio==2.4.1
importlib-resources==3.3.1
joblib==0.14.1
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.1.1
matplotlib==2.2.5
mock==3.0.5
moviepy==1.0.3
mxnet-cu80==1.5.0
networkx==2.2
numba==0.47.0
numpy==1.16.6
opencv-contrib-python==4.2.0.32
pandas==0.24.2
pathlib2==2.3.7.post1
Pillow==6.2.2
proglog==0.1.10
protobuf==3.8.0
PyAudio==0.2.11
pycparser==2.21
pyparsing==2.4.7
python-dateutil==2.8.2
python-speech-features==0.6
pytz==2022.1
PyWavelets==1.0.3
PyYAML==5.4.1
requests==2.27.1
resampy==0.2.2
scandir==1.10.0
scikit-image==0.14.5
scikit-learn==0.20.4
scipy==1.0.1
singledispatch==3.7.0
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
subprocess32==3.5.4
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
Theano==1.0.5
torch @ https://download.pytorch.org/whl/cu80/torch-0.4.0-cp27-cp27mu-linux_x86_64.whl
tqdm==4.64.0
typing==3.10.0.0
urllib3==1.26.9
Werkzeug==1.0.1
wrapt==1.14.1
xlrd==1.2.0
zipp==1.2.0
